Move things into new ParserUtils object
